improve(project): Create new hooks called useSiteMetadata
