+++ src/components/index.ts
@@ -1,6 +1,6 @@
-export { Button } from './Button';
-export { Layout } from './Layout';
-export { SEO } from './SEO';
-export { Text } from './Text';
-export { Title, Subtitle } from './Heading';
-export { Wrapper } from './Wrapper';
+export { Button } from './Button'
+export { Layout } from './Layout'
+export { SEO } from './SEO'
+export { Text } from './Text'
+export { Title, Subtitle } from './Heading'
+export { Wrapper } from './Wrapper'
