+++ tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  purge: [],
+  purge: ['./src/**/*.{ts,tsx}'],
   presets: [],
   darkMode: false, // or 'media' or 'class'
   theme: {
